+++ quadratic-core/src/controller/operations/cells.rs
@@ -15,19 +15,14 @@ use crate::{
 use super::operation::Operation;
 
 impl GridController {
-    pub(super) fn string_to_cell_value(
+    pub fn string_to_cell_value(
         &mut self,
         sheet_pos: SheetPos,
         value: &str,
-        ops: &mut Vec<Operation>,
-    ) -> CellValue {
-        let sheet = self.grid.sheet_mut_from_id(sheet_pos.sheet_id);
-        let sheet_rect = SheetRect::from(sheet_pos);
-
-        // strip whitespace
-        let value = value.trim();
-
-        if value.is_empty() {
+    ) -> (Vec<Operation>, CellValue) {
+        let mut ops = vec![];
+        let sheet_rect: SheetRect = sheet_pos.into();
+        let cell_value = if value.is_empty() {
             CellValue::Blank
         } else if let Some((currency, number)) = CellValue::unpack_currency(value) {
             let numeric_format = NumericFormat {
@@ -42,6 +37,7 @@ impl GridController {
                 )),
             });
             // only change decimal places if decimals have not been set
+            let sheet = self.grid.sheet_from_id(sheet_pos.sheet_id);
             if sheet
                 .get_formatting_value::<NumericDecimals>(sheet_pos.into())
                 .is_none()
@@ -66,11 +62,11 @@ impl GridController {
                     1,
                 )),
             });
-
             CellValue::Number(percent)
         } else {
             CellValue::Text(value.into())
-        }
+        };
+        (ops, cell_value)
     }
 
     pub fn set_cell_value_operations(
@@ -79,12 +75,19 @@ impl GridController {
         value: String,
     ) -> Vec<Operation> {
         let mut ops = vec![];
-        let cell_value = self.string_to_cell_value(sheet_pos, value.as_str(), &mut ops);
 
-        vec![Operation::SetCellValues {
+        // strip whitespace
+        let value = value.trim();
+
+        // convert the string to a cell value and generate necessary operations
+        let (operations, cell_value) = self.string_to_cell_value(sheet_pos, value);
+        ops.extend(operations);
+
+        ops.push(Operation::SetCellValues {
             sheet_rect: sheet_pos.into(),
             values: Array::from(cell_value),
-        }]
+        });
+        ops
     }
 
     pub fn set_cell_code_operations(
@@ -188,6 +191,15 @@ impl GridController {
             sheet_rect,
             borders,
         });
+        ops
+    }
+
+    pub fn delete_values_and_formatting_operations(
+        &mut self,
+        sheet_rect: SheetRect,
+    ) -> Vec<Operation> {
+        let mut ops = self.delete_cells_rect_operations(sheet_rect);
+        ops.extend(self.clear_formatting_operations(sheet_rect));
         ops
     }
 }
