+++ quadratic-core/src/controller/user_actions/clipboard.rs
@@ -1,33 +1,13 @@
-use super::{
-    operations::operation::Operation, transaction_in_progress::TransactionType,
-    transaction_summary::TransactionSummary, GridController,
+use super::super::{
+    transaction_in_progress::TransactionType, transaction_summary::TransactionSummary,
+    GridController,
 };
 use crate::{
-    grid::{
-        formatting::CellFmtArray, generate_borders_full, get_cell_borders_in_rect, BorderSelection,
-        CellBorders, CodeCellValue,
-    },
-    Array, ArraySize, CellValue, Pos, SheetPos, SheetRect,
+    controller::operations::clipboard::{Clipboard, ClipboardCell},
+    grid::{get_cell_borders_in_rect, CodeCellValue},
+    Pos, SheetPos, SheetRect,
 };
 use htmlescape;
-use regex::Regex;
-use serde::{Deserialize, Serialize};
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct ClipboardCell {
-    pub value: Option<CellValue>,
-    pub spill: Option<CellValue>,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct Clipboard {
-    w: u32,
-    h: u32,
-    cells: Vec<ClipboardCell>,
-    formats: Vec<CellFmtArray>,
-    borders: Vec<(i64, i64, Option<CellBorders>)>,
-    code: Vec<(Pos, CodeCellValue)>,
-}
 
 impl GridController {
     pub fn copy_to_clipboard(&self, sheet_rect: SheetRect) -> (String, String) {
@@ -143,230 +123,9 @@ impl GridController {
         sheet_rect: SheetRect,
         cursor: Option<String>,
     ) -> (TransactionSummary, String, String) {
-        let copy = self.copy_to_clipboard(sheet_rect);
-        let summary = self.delete_values_and_formatting(sheet_rect, cursor);
-        (summary, copy.0, copy.1)
-    }
-
-    fn array_from_clipboard_cells(clipboard: Clipboard) -> Option<Array> {
-        if clipboard.w == 0 && clipboard.h == 0 {
-            return None;
-        }
-
-        let mut array = Array::new_empty(ArraySize::new(clipboard.w, clipboard.h).unwrap());
-        let mut x = 0;
-        let mut y = 0;
-
-        clipboard.cells.iter().for_each(|cell| {
-            let value = cell.value.to_owned().map_or(CellValue::Blank, |v| v);
-            // ignore result errors
-            let _ = array.set(x, y, value);
-
-            x += 1;
-
-            if x == clipboard.w {
-                x = 0;
-                y += 1;
-            }
-        });
-
-        Some(array)
-    }
-
-    fn set_clipboard_cells(
-        &mut self,
-        start_pos: SheetPos,
-        clipboard: Clipboard,
-        cursor: Option<String>,
-    ) -> TransactionSummary {
-        let mut compute = false;
-        let sheet_rect = SheetRect {
-            min: start_pos.into(),
-            max: Pos {
-                x: start_pos.x + (clipboard.w as i64) - 1,
-                y: start_pos.y + (clipboard.h as i64) - 1,
-            },
-            sheet_id: start_pos.sheet_id,
-        };
-        let formats = clipboard.formats.clone();
-        let borders = clipboard.borders.clone();
-        let code = clipboard.code.clone();
-
-        let mut ops = vec![];
-        let values = GridController::array_from_clipboard_cells(clipboard);
-        if let Some(values) = values {
-            ops.push(Operation::SetCellValues { sheet_rect, values });
-            compute = true;
-        }
-
-        let sheet = self.grid.sheet_mut_from_id(start_pos.sheet_id);
-
-        // remove any overlapping code cells (which will automatically set the reverse operations)
-        for x in sheet_rect.x_range() {
-            for y in sheet_rect.y_range() {
-                let pos = Pos { x, y };
-                if let Some(_code_cell) = sheet.get_code_cell(pos) {
-                    // no need to clear code cells that are being pasted
-                    if !code.iter().any(|(code_pos, _)| {
-                        Pos {
-                            x: code_pos.x + start_pos.x,
-                            y: code_pos.y + start_pos.y,
-                        } == pos
-                    }) {
-                        ops.push(Operation::SetCellCode {
-                            sheet_pos: pos.to_sheet_pos(start_pos.sheet_id),
-                            code_cell_value: None,
-                        });
-                        compute = true;
-                    }
-                }
-            }
-        }
-
-        // add copied code cells to the sheet
-        code.iter().for_each(|entry| {
-            let sheet_pos = SheetPos {
-                x: entry.0.x + start_pos.x,
-                y: entry.0.y + start_pos.y,
-                sheet_id: start_pos.sheet_id,
-            };
-            ops.push(Operation::SetCellCode {
-                sheet_pos,
-                code_cell_value: Some(entry.1.clone()),
-            });
-            compute = true;
-        });
-
-        formats.iter().for_each(|format| {
-            ops.push(Operation::SetCellFormats {
-                sheet_rect,
-                attr: format.clone(),
-            });
-        });
-
-        // add borders to the sheet
-        borders.iter().for_each(|(x, y, cell_borders)| {
-            if let Some(cell_borders) = cell_borders {
-                let mut border_selections = vec![];
-                let mut border_styles = vec![];
-                let sheet_rect: SheetRect = SheetPos {
-                    sheet_id: sheet.id,
-                    x: *x + start_pos.x,
-                    y: *y + start_pos.y,
-                }
-                .into();
-
-                cell_borders
-                    .borders
-                    .iter()
-                    .enumerate()
-                    .for_each(|(index, border_style)| {
-                        if let Some(border_style) = border_style.to_owned() {
-                            let border_selection = match index {
-                                0 => BorderSelection::Left,
-                                1 => BorderSelection::Top,
-                                2 => BorderSelection::Right,
-                                3 => BorderSelection::Bottom,
-                                _ => BorderSelection::Clear,
-                            };
-                            border_selections.push(border_selection);
-                            border_styles.push(Some(border_style));
-                        }
-                    });
-
-                let borders = generate_borders_full(
-                    sheet,
-                    &sheet_rect.into(),
-                    border_selections,
-                    border_styles,
-                );
-                ops.push(Operation::SetBorders {
-                    sheet_rect,
-                    borders,
-                });
-            }
-        });
-
-        self.set_in_progress_transaction(ops, cursor, compute, TransactionType::Normal)
-    }
-
-    fn paste_plain_text(
-        &mut self,
-        start_pos: SheetPos,
-        clipboard: String,
-        cursor: Option<String>,
-    ) -> TransactionSummary {
-        let lines: Vec<&str> = clipboard.split('\n').collect();
-
-        let mut operations = vec![];
-
-        let cell_values = lines
-            .iter()
-            .enumerate()
-            .map(|(x, line)| {
-                line.split('\t')
-                    .enumerate()
-                    .map(|(y, value)| {
-                        self.string_to_cell_value(
-                            SheetPos {
-                                x: start_pos.x + x as i64,
-                                y: start_pos.y + y as i64,
-                                sheet_id: start_pos.sheet_id,
-                            },
-                            value,
-                            &mut operations,
-                        )
-                    })
-                    .collect::<Vec<CellValue>>()
-            })
-            .collect::<Vec<Vec<CellValue>>>();
-
-        let array = Array::from(cell_values);
-
-        let sheet_rect = SheetRect::new_pos_span(
-            start_pos.into(),
-            (
-                start_pos.x + array.width() as i64 - 1,
-                start_pos.y + array.height() as i64 - 1,
-            )
-                .into(),
-            start_pos.sheet_id,
-        );
-        operations.push(Operation::SetCellValues {
-            sheet_rect,
-            values: array,
-        });
-
-        self.set_in_progress_transaction(operations, cursor, true, TransactionType::Normal)
-    }
-
-    // todo: parse table structure to provide better pasting experience from other spreadsheets
-    fn paste_html(
-        &mut self,
-        sheet_pos: SheetPos,
-        html: String,
-        cursor: Option<String>,
-    ) -> Result<TransactionSummary, ()> {
-        // use regex to find data-quadratic
-        let re = Regex::new(r#"data-quadratic="(.*)"><tbody"#).unwrap();
-        let Some(data) = re.captures(&html) else {
-            return Err(());
-        };
-        let result = &data.get(1).map_or("", |m| m.as_str());
-
-        // decode html in attribute
-        let unencoded = htmlescape::decode_html(result);
-        if unencoded.is_err() {
-            return Err(());
-        }
-
-        // parse into Clipboard
-        let parsed = serde_json::from_str::<Clipboard>(&unencoded.unwrap());
-        if parsed.is_err() {
-            return Err(());
-        }
-        let clipboard = parsed.unwrap();
-        Ok(self.set_clipboard_cells(sheet_pos, clipboard, cursor))
+        let (ops, plain_text, html) = self.cut_to_clipboard_operations(sheet_rect);
+        let summary = self.set_in_progress_transaction(ops, cursor, false, TransactionType::Normal);
+        (summary, plain_text, html)
     }
 
     pub fn paste_from_clipboard(
@@ -378,18 +137,20 @@ impl GridController {
     ) -> TransactionSummary {
         // first try html
         if let Some(html) = html {
-            let pasted_html = self.paste_html(sheet_pos, html, cursor.clone());
-            if let Ok(pasted_html) = pasted_html {
-                return pasted_html;
+            if let Ok(ops) = self.paste_html_operations(sheet_pos, html) {
+                self.set_in_progress_transaction(ops, cursor, true, TransactionType::Normal)
+            } else {
+                TransactionSummary::default()
             }
         }
-
         // if not quadratic html, then use the plain text
         // first try html
-        if let Some(plain_text) = plain_text {
-            return self.paste_plain_text(sheet_pos, plain_text, cursor);
+        else if let Some(plain_text) = plain_text {
+            let ops = self.paste_plain_text_operations(sheet_pos, plain_text);
+            self.set_in_progress_transaction(ops, cursor, true, TransactionType::Normal)
+        } else {
+            TransactionSummary::default()
         }
-        TransactionSummary::default()
     }
 }
 
@@ -707,13 +468,14 @@ mod test {
             Pos { x: 4, y: 4 },
             sheet_id,
         ));
-        let _ = gc.paste_html(
+        let _ = gc.paste_from_clipboard(
             SheetPos {
                 x: 0,
                 y: 10,
                 sheet_id,
             },
-            html,
+            None,
+            Some(html),
             None,
         );
