+++ quadratic-core/src/controller/user_actions/mod.rs
@@ -3,6 +3,7 @@
 pub mod auto_complete;
 pub mod borders;
 pub mod cells;
+pub mod clipboard;
 pub mod formatting;
 pub mod import;
 pub mod undo;
